Move both cookie related functions to its own folder
