+++ projects/packages/my-jetpack/_inc/hooks/use-notification-watcher/use-backup-needs-attention-notice.tsx
@@ -5,8 +5,8 @@ import { useContext, useEffect, useCallback } from 'react';
 import { NOTICE_PRIORITY_HIGH } from '../../context/constants';
 import { NoticeContext } from '../../context/notices/noticeContext';
 import { applyTimezone } from '../../utils/apply-timezone';
-import checkForCookie from '../../utils/check-for-cookie';
-import createCookie from '../../utils/create-cookie';
+import checkForCookie from '../../utils/cookies/check-for-cookie';
+import createCookie from '../../utils/cookies/create-cookie';
 import preventWidows from '../../utils/prevent-widows';
 import useAnalytics from '../use-analytics';
 import { useGetReadableFailedBackupReason } from './use-get-readable-failed-backup-reason';
