+++ projects/packages/my-jetpack/_inc/hooks/use-notification-watcher/use-expiring-plans-notice.tsx
@@ -2,8 +2,8 @@ import { __ } from '@wordpress/i18n';
 import { useContext, useEffect, useCallback, useMemo } from 'react';
 import { NOTICE_PRIORITY_MEDIUM } from '../../context/constants';
 import { NoticeContext } from '../../context/notices/noticeContext';
-import checkForCookie from '../../utils/check-for-cookie';
-import createCookie from '../../utils/create-cookie';
+import checkForCookie from '../../utils/cookies/check-for-cookie';
+import createCookie from '../../utils/cookies/create-cookie';
 import useAnalytics from '../use-analytics';
 import { useGetExpiringNoticeContent } from './use-get-expiring-notice-content';
 import type { NoticeHookType } from './types';
