+++ projects/packages/my-jetpack/_inc/hooks/use-notification-watcher/use-paid-plan-needs-plugin-install-activation-notice.tsx
@@ -10,8 +10,8 @@ import useInstallPlugins from '../../data/products/use-install-plugins';
 import useSimpleQuery from '../../data/use-simple-query';
 import { getMyJetpackWindowInitialState } from '../../data/utils/get-my-jetpack-window-state';
 import useMyJetpackConnection from '../../hooks/use-my-jetpack-connection';
-import checkForCookie from '../../utils/check-for-cookie';
-import createCookie from '../../utils/create-cookie';
+import checkForCookie from '../../utils/cookies/check-for-cookie';
+import createCookie from '../../utils/cookies/create-cookie';
 import useAnalytics from '../use-analytics';
 import { useGetPaidPlanNeedsPluginsContent } from './get-paid-plan-needs-plugins-content';
 import type { NoticeHookType } from './types';
