+++ projects/packages/my-jetpack/_inc/hooks/use-notification-watcher/use-protect-threats-detected-notice.tsx
@@ -4,8 +4,8 @@ import { useContext, useEffect, useCallback } from 'react';
 import { NOTICE_PRIORITY_MEDIUM } from '../../context/constants';
 import { NoticeContext } from '../../context/notices/noticeContext';
 import useProduct from '../../data/products/use-product';
-import checkForCookie from '../../utils/check-for-cookie';
-import createCookie from '../../utils/create-cookie';
+import checkForCookie from '../../utils/cookies/check-for-cookie';
+import createCookie from '../../utils/cookies/create-cookie';
 import preventWidows from '../../utils/prevent-widows';
 import useAnalytics from '../use-analytics';
 import type { NoticeHookType } from './types';
