+++ projects/packages/my-jetpack/_inc/utils/cookies/check-for-cookie.ts
